refactor: move keda.py to standup/lib/
