+++ llmdbenchmark/standup/steps/step_03_workload_monitoring.py
@@ -12,7 +12,7 @@ from llmdbenchmark.executor.command import CommandExecutor
 from llmdbenchmark.parser.cluster_resource_resolver import effective_accelerator_count
 from llmdbenchmark.standup import wva as wva_mod
 from llmdbenchmark.standup import keda_saturation as keda_sat_mod
-from llmdbenchmark.standup import keda as keda_mod
+from llmdbenchmark.standup.lib import keda as keda_mod
 from llmdbenchmark.utilities.capacity_validator import run_capacity_planner
